+++ sdk_core/comm/define.h
@@ -95,22 +95,23 @@ typedef enum {
   
   kCommandIDLidarWorkModeControl = 0x0100,
   kCommandIDLidarGetInternalInfo = 0x0101,
-  kCommandIDLidarPushMsg = 0x0102,
+  kCommandIDLidarPushMsg         = 0x0102,
 
-  kCommandIDLidarRebootDevice = 0x0200,
-  kCommandIDLidarResetDevice = 0x0201,
-  kCommandIDLidarSetPPSSync = 0x0202,
+  kCommandIDLidarRebootDevice    = 0x0200,
+  kCommandIDLidarResetDevice     = 0x0201,
+  kCommandIDLidarSetPPSSync      = 0x0202,
 
-  kCommandIDLidarPushLog = 0x0300,
-  kCommandIDLidarCollectionLog = 0x0301,
-  kCommandIDLidarLogSysTimeSync = 0x0302,
+  kCommandIDLidarPushLog                = 0x0300,
+  kCommandIDLidarCollectionLog          = 0x0301,
+  kCommandIDLidarLogSysTimeSync         = 0x0302,
+  kCommandIDLidarDebugPointCloudControl = 0x0303,
 
 
-  kCommandIDGeneralRequestUpgrade = 0x0400,
-  kCommandIDGeneralXferFirmware = 0x0401,
-  kCommandIDGeneralCompleteXferFirmware = 0x0402,
+  kCommandIDGeneralRequestUpgrade         = 0x0400,
+  kCommandIDGeneralXferFirmware           = 0x0401,
+  kCommandIDGeneralCompleteXferFirmware   = 0x0402,
   kCommandIDGeneralRequestUpgradeProgress = 0x0403,
-  kCommandIDGeneralRequestFirmwareInfo = 0xFF,
+  kCommandIDGeneralRequestFirmwareInfo    = 0xFF,
   kCommandIDLidarCommandCount
 } LidarCommandID;
 
@@ -171,36 +172,38 @@ typedef struct {
   uint16_t lidar_port;
 } HostIpInfoValue;
 
-static const uint16_t kDetectionPort = 56000;
-static const uint16_t kDetectionListenPort = 56001;
+static const uint16_t kDetectionPort           = 56000;
+static const uint16_t kDetectionListenPort     = 56001;
+static const uint16_t kHostDebugPointCloudPort = 44332;
 
-static const uint16_t kHAPCmdPort = 56000;
-static const uint16_t kHAPPushMsgPort = 56000;
+static const uint16_t kHAPCmdPort       = 56000;
+static const uint16_t kHAPPushMsgPort   = 56000;
 static const uint16_t kHAPPointDataPort = 57000;
-static const uint16_t kHAPIMUPort = 58000;
-static const uint16_t kHAPLogPort = 59000;
+static const uint16_t kHAPIMUPort       = 58000;
+static const uint16_t kHAPLogPort       = 59000;
 
 /** kLogPort, which is the log port to be banned. */
 static const uint16_t kLogPort = 0;
 
 static const uint16_t kHAPLidarCmdPort = 56000;
 
-static const uint16_t kMid360LidarCmdPort = 56100;
-static const uint16_t kMid360LidarPushMsgPort = 56200;
-static const uint16_t kMid360LidarPointCloudPort = 56300;
-static const uint16_t kMid360LidarImuDataPort = 56400;
-static const uint16_t kMid360LidarLogPort = 56500;
+static const uint16_t kMid360LidarCmdPort             = 56100;
+static const uint16_t kMid360LidarPushMsgPort         = 56200;
+static const uint16_t kMid360LidarPointCloudPort      = 56300;
+static const uint16_t kMid360LidarImuDataPort         = 56400;
+static const uint16_t kMid360LidarLogPort             = 56500;
+static const uint16_t kMid360LidarDebugPointCloudPort = 60301;
 
-static const uint16_t kMid360HostCmdPort = 56101;
-static const uint16_t kMid360HostPushMsgPort = 56201;
+static const uint16_t kMid360HostCmdPort        = 56101;
+static const uint16_t kMid360HostPushMsgPort    = 56201;
 static const uint16_t kMid360HostPointCloudPort = 56301;
-static const uint16_t kMid360HostImuDataPort = 56401;
-static const uint16_t kMid360HostLogPort = 56501;
+static const uint16_t kMid360HostImuDataPort    = 56401;
+static const uint16_t kMid360HostLogPort        = 56501;
 
-static const uint16_t kPaLidarCmdPort = 9347;
+static const uint16_t kPaLidarCmdPort        = 9347;
 static const uint16_t kPaLidarPointCloudPort = 10000;
-static const uint16_t kPaLidarFaultPort = 10001;
-static const uint16_t kPaLidarLogPort = 1002;
+static const uint16_t kPaLidarFaultPort      = 10001;
+static const uint16_t kPaLidarLogPort        = 1002;
 
 static const uint16_t kPaHostFaultPort = 42867;
 
@@ -225,6 +228,13 @@ typedef struct {
   uint8_t log_type;
   uint8_t enable;
 } EnableDeviceLoggerRequest;
+
+typedef struct {
+  std::string   sn;
+  std::uint8_t  dev_type;
+  std::string   lidar_ip;
+  std::uint16_t cmd_port;
+} LidarDeviceInfo;
 
 typedef struct {
   uint8_t log_type;         // 0
@@ -310,6 +320,22 @@ typedef struct {
   uint16_t length;  /**< The length of firmware info string, include '\0'. */
   uint8_t info[1];  /**< Firmware info string, include '\0'. */
 } LivoxLidarRequestFirmwareInfoResponse;
+
+typedef struct {
+  uint8_t file_ver;   /**< file format version. */
+  uint8_t dev_type;   /**< the device type of the firmware. */
+  uint8_t data_type;  /**< type of data. */
+  uint8_t sn[16];
+  uint8_t rsvd[107];
+  uint16_t crc16;
+} LivoxLidarDebugPointCloudFileHeader;
+
+typedef struct {
+  uint8_t   enable;
+  uint8_t   host_ip_addr[4];
+  uint16_t  host_port;
+  uint16_t  bandwidth;
+} LivoxLidarDebugPointCloudRequest;
 
 typedef void(*LivoxLidarStartUpgradeCallback)(livox_status status, uint32_t handle,
     LivoxLidarStartUpgradeResponse* response, void* client_data);