+++ sdk_core/command_handler/command_impl.h
@@ -106,6 +106,8 @@ class CommandImpl {
 
   static livox_status LivoxLidarRequestReset(uint32_t handle, LivoxLidarResetCallback cb, void* client_data);
 
+  static livox_status SetLivoxLidarDebugPointCloud(uint32_t handle, bool enable, LivoxLidarLoggerCallback cb, void* client_data);
+
   /*******Upgrade Module***********/    
   static livox_status LivoxLidarRequestReboot(uint32_t handle, LivoxLidarRebootCallback cb, void* client_data);